[ Mac ] imported/w3c/web-platform-tests/web-animations/timing-model/timelines/update-and-send-events-replacement.html is flaky failing. https://bugs.webkit.org/show_bug.cgi?id=209239 <rdar://problem/60591358>
Patch by Antoine Quint <graouts@apple.com> on 2020-03-22
Reviewed by Simon Fraser.

Source/WebCore:

This test was made flaky by r257417, the initial fix for webkit.org/b/208069. A new, appropriate fix for that bug is in the works. In the
meantime we revert r257417 in this patch.

The reason this test became flaky is that it features the following code:

    animB.timeline = new DocumentTimeline({
      originTime:
        document.timeline.currentTime - 100 * MS_PER_SEC - animB.startTime,
    });

In this case the only reference to the created DocumentTimeline is through `animB.timeline`. But because r257417 made the timeline reference from
WebAnimation a weak reference, in some cases, if GC kicks in, the timeline would be dereferenced and the test would fail. We restore that relationship
to its previous state, which is a strong reference.

* animation/WebAnimation.cpp:
(WebCore::WebAnimation::setTimeline):
(WebCore::WebAnimation::setTimelineInternal):
(WebCore::WebAnimation::enqueueAnimationEvent):
(WebCore::WebAnimation::acceleratedStateDidChange):
(WebCore::WebAnimation::timeline const): Deleted.
* animation/WebAnimation.h:
(WebCore::WebAnimation::timeline const):

LayoutTests:

* platform/mac/TestExpectations:

Identifier: 222328@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@258823 268f45cc-cd09-0410-ab3c-d52691b4dbfc