AudioMediaStreamTrackRendererCocoa does not recover from AudioSession interruption https://bugs.webkit.org/show_bug.cgi?id=209412
Reviewed by Eric Carlson.

In case there is an AudioSession interruption, like the app is in the background
and another app starts to play audio, AudioMediaStreamTrackRendererCocoa will not restart playing audio.
Fix this by clearing the AudioUnit when starting since the renderer is paused during the interruption.
This ensures resuming from the interruption in a clean state.

Manually tested.

* platform/mediastream/mac/AudioMediaStreamTrackRendererCocoa.cpp:
(WebCore::AudioMediaStreamTrackRendererCocoa::start):


Identifier: 222344@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@258839 268f45cc-cd09-0410-ab3c-d52691b4dbfc