[Flatpak SDK] Pass all the arguments of build-webkit to webkit-flatpak https://bugs.webkit.org/show_bug.cgi?id=209558
Reviewed by Žan Doberšek.

When using flatpak some of the arguments we pass to build-webkit
are not meant to be used by that script but by webkit-flatpak. However we are
not passing all of them to webkit-flatpak but just the configuration ones
(port, release/debug...). This means that all the arguments that configure the
behaviour of webkit-flatpak are lost.

* Scripts/webkitdirs.pm:
(runInFlatpak): Filter-out Flatpak SDK-specific arguments to a
separate array, passed to webkit-flatpak.

Identifier: 222523@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@259045 268f45cc-cd09-0410-ab3c-d52691b4dbfc