Unreviewed, fix the watchOS build after r259151
The declaration of Pasteboard(const String&) in Pasteboard.h is present for all PLATFORM(IOS_FAMILY), but the
implementation is guarded by ENABLE(DRAG_SUPPORT). r259151 added a codepath that calls this constructor in
IOS_FAMILY code, causing a linker error. Fix this by moving the implementation out of the ENABLE(DRAG_SUPPORT)
guard, to match the declaration in the header.

* platform/ios/PasteboardIOS.mm:

Identifier: 222628@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@259163 268f45cc-cd09-0410-ab3c-d52691b4dbfc