Division by zero in RenderBlockFlow::computeColumnCountAndWidth https://bugs.webkit.org/show_bug.cgi?id=209485 <rdar://problem/60746109>
Reviewed by Zalan Bujtas.

Source/WebCore:

When computing content width and height, set it to 0 if the computed size
is negative.

Test: fast/multicol/negativeColumnGap.html

* rendering/RenderBox.h:
(WebCore::RenderBox::contentWidth const):
(WebCore::RenderBox::contentHeight const):

LayoutTests:

Added a regression test for the crash. Also modify the expected output
of button.html because the size would now be different.

* fast/multicol/negativeColumnGap-expected.txt: Added.
* fast/multicol/negativeColumnGap.html: Added.
* platform/mac/css3/flexbox/button-expected.txt:


Identifier: 222675@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@259210 268f45cc-cd09-0410-ab3c-d52691b4dbfc