Assertion failure in HTMLFormElement::formElementIndex https://bugs.webkit.org/show_bug.cgi?id=209643
Reviewed by Darin Adler.

The bug was caused by FormAssociatedElement::findAssociatedForm finding a wrong form element
when it's called on an element which appears later in the removed subtree.

When we find the new form element to associate this element with, check to make sure its root
element is that of the tree scope. This condition will be false if this element is in in the midst
of being removed.

* html/FormAssociatedElement.cpp:
(WebCore::FormAssociatedElement::findAssociatedForm):


Identifier: 222700@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@259244 268f45cc-cd09-0410-ab3c-d52691b4dbfc