The relative frame and hit test of isolated objects must be dispatched to the main thread. https://bugs.webkit.org/show_bug.cgi?id=209792
Reviewed by Chris Fleizach.

The relative frame of isolated objects must be calculated on the main
thread because it requires the scroll ancestor to convert to the
appropriate scroll offset. The relative frame cannot be cached because
the scroll offset can change.
Accordingly, the hit test cannot rely on a cached relative frame and
must be dispatched to be computed on the main thread as well.

* accessibility/AXObjectCache.h:
* accessibility/isolatedtree/AXIsolatedObject.cpp:
(WebCore::AXIsolatedObject::initializeAttributeData): Do not cache the relative frame any longer.
(WebCore::AXIsolatedObject::accessibilityHitTest const): Dispatched to the main thread.
(WebCore::AXIsolatedObject::relativeFrame const): Dispatched to the main thread.
* accessibility/isolatedtree/AXIsolatedObject.h:
* accessibility/mac/WebAccessibilityObjectWrapperMac.mm:
(-[WebAccessibilityObjectWrapper position]):
(-[WebAccessibilityObjectWrapper accessibilityAttributeValue:]):
(-[WebAccessibilityObjectWrapper accessibilityHitTest:]):


Identifier: 222723@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@259288 268f45cc-cd09-0410-ab3c-d52691b4dbfc