Allow WKAppBoundDomains to be initialized with eTLD+1 only (no protocol) https://bugs.webkit.org/show_bug.cgi?id=209839 <rdar://problem/61129400>
Reviewed by Darin Adler.

Source/WebKit:

Create a convenience mode for WKAppBoundDomains that assumes https if the user does
not supply the full URL. This doesn't effect the behavior of the app-bound domains
because we only deal in RegistrableDomains.

Tested by TestWebKitAPI.

* UIProcess/WebsiteData/Cocoa/WebsiteDataStoreCocoa.mm:
(WebKit::WebsiteDataStore::initializeAppBoundDomains): If the protocol is missing from
a domain supplied by WKAppBoundDomains, assume it was https.

Tools:

* TestWebKitAPI/Info.plist:


Identifier: 222754@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@259322 268f45cc-cd09-0410-ab3c-d52691b4dbfc