[Web Animations] Move Document.getAnimations() to DocumentOrShadowRoot https://bugs.webkit.org/show_bug.cgi?id=202192 <rdar://problem/55697775>
Reviewed by Antti Koivisto.

LayoutTests/imported/w3c:

Update the test relevant to DocumentOrShadowRoot.getAnimations() from upstream and record three new PASS results in it. We also get two new PASS from a harness test.

* web-platform-tests/web-animations/idlharness.window-expected.txt:
* web-platform-tests/web-animations/interfaces/DocumentOrShadowRoot/getAnimations-expected.txt:
* web-platform-tests/web-animations/interfaces/DocumentOrShadowRoot/getAnimations.html:
* web-platform-tests/web-animations/testcommon.js:
(async insertFrameAndAwaitLoad):

Source/WebCore:

We remove the getAnimations() declaration from the Document interface and instead move it on the DocumentOrShadowRoot interface.

We add the new method Document::matchingAnimations() which takes a lambda that is provided an animation's effect's target to determine whether
that animation should be found in the list of animations.

In the case of Document::getAnimations(), we filter out animations targeting elements hosted in shadow roots, while in ShadowRoot:getAnimations(),
we filter out animations targeting elements that are not hosted in the shadow root the method was called on.

* dom/Document.cpp:
(WebCore::Document::getAnimations):
(WebCore::Document::matchingAnimations):
* dom/Document.h:
* dom/Document.idl:
* dom/DocumentOrShadowRoot.idl:
* dom/Element.cpp:
(WebCore::Element::getAnimations):
* dom/ShadowRoot.cpp:
(WebCore::ShadowRoot::getAnimations):
* dom/ShadowRoot.h:

Identifier: 222968@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@259577 268f45cc-cd09-0410-ab3c-d52691b4dbfc