TextManipulationController fails to replace a paragraph that ends with a br https://bugs.webkit.org/show_bug.cgi?id=210099
Reviewed by Wenson Hsieh.

Source/WebCore:

The bug was caused by TextManipulationController::replace not ignoring the br at the end of a paragraph
even through it doesn't appear as a token. We also need to insert this br back at the end of the paragraph
when completing the manipulation.

* editing/TextManipulationController.cpp:
(WebCore::TextManipulationController::replace):

Tools:

Added regression tests.

* TestWebKitAPI/Tests/WebKitCocoa/TextManipulation.mm:
(TextManipulation.CompleteTextManipulationReplaceMultipleSimpleParagraphsSeparatedByBR):
(TextManipulation.CompleteTextManipulationReplaceParagraphsSeparatedByWrappedBR):
(TextManipulation.CompleteTextManipulationFailWhenBRIsInserted):


Identifier: 223025@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@259647 268f45cc-cd09-0410-ab3c-d52691b4dbfc