Web Inspector: Debugger: treat comma sub-expressions as separate statements https://bugs.webkit.org/show_bug.cgi?id=209998
Reviewed by Joseph Pecoraro.

Source/JavaScriptCore:

Minifiers sometimes use `,` instead of `;` as a way of compressing multiple statements into
a single statement (e.g. inside an `if` conditional). This makes stepping through minified
code much less useful, as our current stepping logic is based on expressions (not including
step in or step out). In most cases, we do actually want to treat each sub-expression in a
comma expression as a statement so as to emulate the behaviour as if there were `;` instead.

* bytecompiler/NodesCodegen.cpp:
(JSC::CommaNode::emitBytecode):

* parser/Parser.cpp:
(JSC::Parser<LexerType>::parseVariableDeclarationList):
(Parser<LexerType>::parseExpression):
Shift the location of subsequent sub-expressions to be after the comma. Add breakpoint pause
locations for each subsequent sub-expression.

Source/WebInspectorUI:

Minifiers sometimes use `,` instead of `;` as a way of compressing multiple statements into
a single statement (e.g. inside an `if` conditional). This makes stepping through minified
code much less useful, as our current stepping logic is based on expressions (not including
step in or step out). In most cases, we do actually want to treat each sub-expression in a
comma expression as a statement so as to emulate the behaviour as if there were `;` instead.

* UserInterface/Workers/Formatter/JSFormatter.js:
(JSFormatter.prototype._isLikelyToHaveNewline):
(JSFormatter.prototype._handleTokenAtNode):
Put each sub-expression in a comma expression on it's own line.

LayoutTests:

* inspector/debugger/breakpoints/resources/dump-functions.js:
* inspector/debugger/breakpoints/resources/dump-general.js:
* inspector/debugger/breakpoints/resolved-dump-all-pause-locations-expected.txt:
* inspector/debugger/breakpoints/resolved-dump-each-line-expected.txt:
* inspector/debugger/stepping/stepOver.html:
* inspector/debugger/stepping/stepOver-expected.txt:
* inspector/debugger/stepping/stepping-function-calls.html:
* inspector/debugger/stepping/stepping-function-calls-expected.txt:
* inspector/debugger/stepping/stepping-loops.html:
* inspector/debugger/stepping/stepping-loops-expected.txt:
* inspector/debugger/stepping/stepping-misc.html:
* inspector/debugger/stepping/stepping-misc-expected.txt:
* inspector/formatting/resources/javascript-tests/for-statements.js:
* inspector/formatting/resources/javascript-tests/for-statements-expected.js:
* inspector/formatting/resources/javascript-tests/if-statements.js:
* inspector/formatting/resources/javascript-tests/if-statements-expected.js:
* inspector/formatting/resources/javascript-tests/sample-jquery-expected.js:
* inspector/formatting/resources/javascript-tests/sample-jquery-expected.js:


Identifier: 223141@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@259781 268f45cc-cd09-0410-ab3c-d52691b4dbfc