Fix for crash in test accessibility/mac/aria-grid-with-strange-hierarchy.html in isolated tree mode. https://bugs.webkit.org/show_bug.cgi?id=210295
Reviewed by Chris Fleizach.

Covered by accessibility/mac/aria-grid-with-strange-hierarchy.html.

- When AXIsolatedTree::applyPendingChanges encounters a change for an
already existing object, the existing object is discarded and the new
object replaces it in the nodes map. The existing and new objects must
have the same platform wrapper. Thus the wrapper needs to be detached
from the existing object about to be discarded, and re-attached to the
new object. We were missing the re-attachment, and hence the crash when
the wrapper tries to access its underlying object.
- In addition, moved the LockHolder in a couple of intances to before
AXIsolatedTree::nodeForID, because this method accesses a member
variable used in both threads.
- Added stricter assert checks to catch problems with the management of
objects and wrappers during tree updates.

* accessibility/isolatedtree/AXIsolatedTree.cpp:
(WebCore::AXIsolatedTree::updateNode):
(WebCore::AXIsolatedTree::updateChildren):
(WebCore::AXIsolatedTree::applyPendingChanges):
* accessibility/mac/WebAccessibilityObjectWrapperBase.mm:
(-[WebAccessibilityObjectWrapperBase attachIsolatedObject:]):


Identifier: 223191@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@259832 268f45cc-cd09-0410-ab3c-d52691b4dbfc