[ Mac wk1 ] fast/images/animated-gif-restored-from-bfcache.html is a flaky failure https://bugs.webkit.org/show_bug.cgi?id=206950 <rdar://problem/58998008>
Reviewed by Alex Christensen.

Make sure the test calls testRunner.waitUntilDone() synchronously when loaded instead
of doing it later on only when returning from the back/forward cache. Also, trigger
the navigation in a setTimeout(0) in the load event and rely on the
page-cache-helper.html like other back/forward cache tests do.

* fast/images/animated-gif-restored-from-bfcache.html:
* platform/mac-wk1/TestExpectations:


Identifier: 223242@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@259888 268f45cc-cd09-0410-ab3c-d52691b4dbfc